fix(site): 移除官网原来的next-sdk, robot的实现
