+++ examples/sites/package.json
@@ -27,11 +27,12 @@
     "@docsearch/css": "^3.8.0",
     "@docsearch/js": "^3.8.0",
     "@docsearch/react": "npm:@docsearch/css",
-    "@opentiny/next-sdk": "0.0.1-alpha.5",
+    "@opentiny/next-remoter": "^0.0.1-alpha.10",
+    "@opentiny/next-sdk": "^0.1.1",
     "@opentiny/tiny-robot": "0.3.0-alpha.3",
     "@opentiny/tiny-robot-kit": "0.3.0-alpha.3",
     "@opentiny/tiny-robot-svgs": "0.3.0-alpha.3",
-    "@opentiny/tiny-vue-mcp": "^0.0.2",
+    "@opentiny/tiny-vue-mcp": "^0.0.3",
     "@opentiny/utils": "workspace:~",
     "@opentiny/vue": "workspace:~",
     "@opentiny/vue-common": "workspace:~",
